// packages/ui/src/components/CheckboxChip/types.ts
import type { CheckboxRootChangeEventDetails } from '@base-ui/react/checkbox';
import type {
  FocusEventHandler,
  KeyboardEventHandler,
  MouseEventHandler,
  ReactNode,
  Ref,
} from 'react';

export type CheckboxChipColor =
  | 'primary'
  | 'secondary'
  | 'info'
  | 'success'
  | 'warning'
  | 'error';

export type CheckboxChipSize = 's' | 'm';

export type CheckboxChipProps = {
  'aria-describedby'?: string;
  'aria-label'?: string;
  'aria-labelledby'?: string;
  checked?: boolean;
  children?: ReactNode;
  color?: CheckboxChipColor;
  defaultChecked?: boolean;
  disabled?: boolean;
  id?: string;
  indeterminate?: boolean;
  inputRef?: Ref<HTMLInputElement>;
  name?: string;
  onBlur?: FocusEventHandler<HTMLElement>;
  onCheckedChange?: (
    checked: boolean,
    details: CheckboxRootChangeEventDetails
  ) => void;
  onClick?: MouseEventHandler<HTMLElement>;
  onFocus?: FocusEventHandler<HTMLElement>;
  onKeyDown?: KeyboardEventHandler<HTMLElement>;
  parent?: boolean;
  readOnly?: boolean;
  ref?: Ref<HTMLElement>;
  required?: boolean;
  size?: CheckboxChipSize;
  tabIndex?: number;
  title?: string;
  uncheckedValue?: string;
  value?: string;
};